Optimize Docker build workflows in Github Actions
The critical change here is pushing the ARMv7 build off to an ARM64 runner. Running the ARMv7 build on
x86_64/amd64 architecture runners requires QEMU emulation of ARM64 which incurs a dramatic reduction in
performance.

Some of the actions have been updated as many were using deprecated versions or unmaintained.